patch tools/query
