+++ tools/query/Cargo.toml
@@ -11,7 +11,8 @@ repository = { workspace = true }
 
 [dependencies]
 zapatos-sdk = { workspace = true }
-libra-config = { workspace = true }
+# libra-config = { workspace = true }
+libra-types = { workspace = true }
 anyhow = { workspace = true }
 async-trait = { workspace = true }
 clap = { workspace = true }
