--- tools/query/src/extension/mod.rs
@@ -1,1 +1,0 @@
-pub mod client_ext;
