+++ tools/query/src/lib.rs
@@ -1,2 +1,2 @@
-pub mod extension;
+
 pub mod querier;
