+++ types/src/type_extensions/client_ext.rs
@@ -35,6 +35,8 @@ pub const USER_AGENT: &str = concat!("libra-config/", env!("CARGO_PKG_VERSION"))
 pub trait ClientExt {
     fn default() -> Result<Client>;
 
+    async fn get_account_resources_ext(&self, account: AccountAddress) -> Result<String>;
+
     async fn get_sequence_number(&self, account: AccountAddress) -> Result<u64>;
 
     async fn generate_transaction(
@@ -73,6 +75,14 @@ impl ClientExt for Client {
         ))
     }
 
+    async fn get_account_resources_ext(&self, account: AccountAddress) -> Result<String> {
+        let response = self
+            .get_account_resources(account)
+            .await
+            .context("Failed to get account resources")?;
+        Ok(format!("{:#?}", response.inner()))
+    }
+    
     async fn get_sequence_number(&self, account: AccountAddress) -> Result<u64> {
         let response = self
             .get_account_resource(account, "0x1::account::Account")
